requirements.txt
pycryptodomex
pandas
tqdm
requests
types-requests
pyyaml
types-PyYAML
jinja2
types-pkg-resources
zstandard
appdirs
texttable
build
pkginfo
boto3>=1.23.0
urllib3<2,>=1.21.1
requests==2.31.0
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
streamlit==1.29.0
streamlit_image-coordinates==0.1.6
streamlit_dimensions==0.0.1